requirements.txt
asgiref==3.7.2
colorama==0.4.6
coverage==7.13.1
Django==4.2.7
django-jazzmin==2.6.0
django-simpleui==2025.6.24
djangorestframework==3.16.1
factory_boy==3.3.3
Faker==40.1.0
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
